Move ProviderKind into metadata module
