+++ crates/bifrost/src/loglet.rs
@@ -11,38 +11,11 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use enum_map::Enum;
 
 use restate_types::logs::{Lsn, Payload, SequenceNumber};
 
-use crate::metadata::LogletParams;
+use crate::metadata::{LogletParams, ProviderKind};
 use crate::{Error, LogRecord, LsnExt, Options, ProviderError};
-
-/// An enum with the list of supported loglet providers.
-/// For each variant we must have a corresponding implementation of the
-/// [`crate::loglet::Loglet`] trait
-#[derive(
-    Debug,
-    Clone,
-    Hash,
-    Eq,
-    PartialEq,
-    Copy,
-    serde::Serialize,
-    serde::Deserialize,
-    Enum,
-    strum_macros::EnumIter,
-    strum_macros::Display,
-)]
-#[serde(rename_all = "snake_case")]
-pub enum ProviderKind {
-    #[cfg(any(test, feature = "local_loglet"))]
-    /// A local rocksdb-backed loglet.
-    Local,
-    #[cfg(any(test, feature = "memory_loglet"))]
-    /// An in-memory loglet, primarily for testing.
-    InMemory,
-}
 
 // why? because if all loglet features are disabled, clippy will complain about options being
 // unused.
